add a new file config/readableNames for dictionaries used in helpers/getReadable (extractd from the helper file)
